Change parsing to work inside any monad
This lays the groundwork for adding environment-dependent parse
elements. For things like a "last result" symbol, there are two main
options: adding a special expression symbol or modifying the parser to
reference old results.

Adding a special expression symbol isn't a good idea, since it'd require
changes to everything related to approximation and rewriting. In some
contexts, this symbol wouldn't make any sense - if there's no past
results or if the expression isn't being evaluated in a REPL context,
there wouldn't be anything there to reference.

This commit prepares the expression parser to operate inside EnvM.