Add wrapper type for name references
This commit wraps variable names in a reference type in the context of
expressions. It's intended to prepare for unit assertions for variables,
so rather than "2N/x" you could write "2N/ x:m/s/s", which asserts that
the value of x must be dimensionally compatible with "m/s/s".

Adding this will also allow automatic unit conversions when substituting
values for variables with dimensionally compatible types later.