Account for units during reduction grouping
This commit fixes a failure to reduce when given a constant expression
with globally incompatible units, but which can be partially reduced.
For example "2N-4kg+8kg". This should be partially reduced to its
simpler form "2N+4kg", since the kilogram values can be combined.

Previously, the equation would not be reduced since reduction did not
account for units when reordering addition/subtraction chains. This
commit fixes the issue by taking units into account.