Fix decimal approximation bug in subtract
This commit fixes a decimal approximation bug when subtracting exact
reals. The expression "1.199 - 1" would produce "1.202" rather than the
correct result, ".199". This occurred because the implementation of
subtract did not rescale its arguments properly and was actually
*adding* the parameters rather than subtracting them.

This fixes the issue by turning the approximate subtract implementation
so it just delegates to addition. This should prevent any similar bugs
in the future since the behavior of both addition and subtraction are in
the same place.