Combine orders of magnitude when reducing units
This commit adds order-of-magnitude changes while performing
multiplication and division on units. For example, before this commit
the expression "1km / 1m" would not produce "1" as it should. "1GJ / 1W"
would also not produce "1Gs" like it's supposed to.

This fixes the issue by adding a new reduction function that takes
orders of magnitude into account. Since it's slower, this function is
only run when approximating a result, not when combining units normally.