Fix exact real rendering
There was an issue when rendering an exact real that had been converted
to reduced form. Specifically, when the number of digits in the rendered
base was less than the magnitude of the exponent.

Decimal values were particularly affected. 1*10^-6 was rendering as ".1"
rather than as ".000001" as it should. This commit fixes the issue by
changing the exact real rendering algorithm to account for these cases.